fix airspy_adsb service not stopping correctly (#23)
The 'trap ... pkill ...' in the service script doesn't work, making s6
wait for the service indefinitely.  This in turn runs into the 10 second
docker stop / compose down timeout which will send a SIGKILL to the
whole container leading to ungraceful termination.

The trap will only work when there is no foreground program being
executed by the shell, it does work however during a wait.

Start airspy_adsb in the background and issue wait.
Run the sleep delay in the background as well so we don't need to wait
for it when the trap pkill happens. Should airspy_adsb error, the wait
should limit program starts to every 10 seconds.